LPS-56029 Extract AuthVerifier functionalities from AutoLogins
