+++ modules/portal-security/portal-security-auth/src/com/liferay/portal/security/auth/verifier/BasicAuthHeaderAuthVerifier.java
@@ -1,0 +1,108 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.auth.verifier;
+
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.security.auth.verifier.AuthVerifier;
+import com.liferay.portal.kernel.security.auth.verifier.AuthVerifierResult;
+import com.liferay.portal.kernel.security.auto.login.AutoLoginException;
+import com.liferay.portal.kernel.servlet.HttpHeaders;
+import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.security.auth.AccessControlContext;
+import com.liferay.portal.security.auth.AuthException;
+import com.liferay.portal.security.auto.login.BasicAuthHeaderAutoLogin;
+import com.liferay.portal.util.Portal;
+
+import java.util.Properties;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Tomas Polesovsky
+ */
+@Component(
+	immediate = true,
+	property = {
+		"auth.verifier.BasicAuthHeaderAuthVerifier.basic_auth=false",
+		"auth.verifier.BasicAuthHeaderAuthVerifier.hosts.allowed=",
+		"auth.verifier.BasicAuthHeaderAuthVerifier.urls.excludes=/api/liferay/*",
+		"auth.verifier.BasicAuthHeaderAuthVerifier.urls.includes=/api/*,/xmlrpc/*"
+	},
+	service = AuthVerifier.class
+)
+public class BasicAuthHeaderAuthVerifier
+	extends BasicAuthHeaderAutoLogin implements AuthVerifier {
+
+	@Override
+	public String getAuthType() {
+		return HttpServletRequest.BASIC_AUTH;
+	}
+
+	@Override
+	public AuthVerifierResult verify(
+			AccessControlContext accessControlContext, Properties properties)
+		throws AuthException {
+
+		try {
+			AuthVerifierResult authVerifierResult = new AuthVerifierResult();
+
+			String[] credentials = login(
+				accessControlContext.getRequest(),
+				accessControlContext.getResponse());
+
+			if (credentials != null) {
+				authVerifierResult.setPassword(credentials[1]);
+				authVerifierResult.setState(AuthVerifierResult.State.SUCCESS);
+				authVerifierResult.setUserId(Long.valueOf(credentials[0]));
+			}
+			else {
+
+				// Deprecated
+
+				boolean forcedBasicAuth = MapUtil.getBoolean(
+					accessControlContext.getSettings(), "basic_auth");
+
+				if (forcedBasicAuth) {
+					HttpServletResponse response =
+						accessControlContext.getResponse();
+
+					response.setHeader(
+						HttpHeaders.WWW_AUTHENTICATE, _BASIC_REALM);
+
+					response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+
+					authVerifierResult.setState(
+						AuthVerifierResult.State.INVALID_CREDENTIALS);
+				}
+			}
+
+			return authVerifierResult;
+		}
+		catch (AutoLoginException ale) {
+			throw new AuthException(ale);
+		}
+	}
+
+	private static final String _BASIC_REALM =
+		"Basic realm=\"" + Portal.PORTAL_REALM + "\"";
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BasicAuthHeaderAuthVerifier.class);
+
+}
