+++ modules/portal-security/portal-security-auth/src/com/liferay/portal/security/auth/verifier/ParameterAuthVerifier.java
@@ -1,0 +1,73 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.auth.verifier;
+
+import com.liferay.portal.kernel.security.auth.verifier.AuthVerifier;
+import com.liferay.portal.kernel.security.auth.verifier.AuthVerifierResult;
+import com.liferay.portal.kernel.security.auto.login.AutoLoginException;
+import com.liferay.portal.security.auth.AccessControlContext;
+import com.liferay.portal.security.auth.AuthException;
+import com.liferay.portal.security.auto.login.ParameterAutoLogin;
+
+import java.util.Properties;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Tomas Polesovsky
+ */
+@Component(
+	immediate = true,
+	property = {
+		"auth.verifier.ParameterAuthVerifier.hosts.allowed=255.255.255.255",
+		"auth.verifier.ParameterAuthVerifier.urls.excludes=*",
+		"auth.verifier.ParameterAuthVerifier.urls.includes="
+	},
+	service = AuthVerifier.class
+)
+public class ParameterAuthVerifier
+	extends ParameterAutoLogin implements AuthVerifier {
+
+	@Override
+	public String getAuthType() {
+		return getClass().getSimpleName();
+	}
+
+	@Override
+	public AuthVerifierResult verify(
+			AccessControlContext accessControlContext, Properties properties)
+		throws AuthException {
+
+		try {
+			AuthVerifierResult authVerifierResult = new AuthVerifierResult();
+
+			String[] credentials = login(
+				accessControlContext.getRequest(),
+				accessControlContext.getResponse());
+
+			if (credentials != null) {
+				authVerifierResult.setPassword(credentials[1]);
+				authVerifierResult.setState(AuthVerifierResult.State.SUCCESS);
+				authVerifierResult.setUserId(Long.valueOf(credentials[0]));
+			}
+
+			return authVerifierResult;
+		}
+		catch (AutoLoginException ale) {
+			throw new AuthException(ale);
+		}
+	}
+
+}
