+++ portal-impl/src/portal.properties
@@ -3611,15 +3611,15 @@
     # This property is not read by the portal except for portal properties
     # overridden by liferay-hook.xml. It remains here only as a reference.
     #
-    #auth.verifier.pipeline=com.liferay.portal.security.auto.login.BasicAuthHeaderAutoLogin,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auto.login.ParameterAutoLogin,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
+    #auth.verifier.pipeline=com.liferay.portal.security.auth.verifier.BasicAuthHeaderAuthVerifier,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auth.verifier.ParameterAuthVerifier,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
 
     #
     # Each authentication verifier can have custom properties set via the
     # property prefix "auth.verifier." and the authentication verifier's simple
     # class name. All property suffixes are stripped of their prefix and passed
     # directly to the authentication verifier. For example, the property
-    # "auth.verifier.BasicAuthHeaderAutoLogin.hosts.allowed" is passed to
-    # BasicAuthHeaderAutoLogin as "hosts.allowed".
+    # "auth.verifier.BasicAuthHeaderAuthVerifier.hosts.allowed" is passed to
+    # BasicAuthHeaderAuthVerifier as "hosts.allowed".
     #
     # The expected property suffixes are "hosts.allowed" and "urls". See the
     # property "json.service.auth.token.hosts.allowed" for the accepted values
@@ -3633,11 +3633,11 @@
     #
 
     #
-    # BasicAuthHeaderAutoLogin
+    # BasicAuthHeaderAuthVerifier
     #
-    #auth.verifier.BasicAuthHeaderAutoLogin.hosts.allowed=
-    #auth.verifier.BasicAuthHeaderAutoLogin.urls.includes=/api/*,/xmlrpc/*
-    #auth.verifier.BasicAuthHeaderAutoLogin.urls.excludes=/api/liferay/do
+    #auth.verifier.BasicAuthHeaderAuthVerifier.hosts.allowed=
+    #auth.verifier.BasicAuthHeaderAuthVerifier.urls.includes=/api/*,/xmlrpc/*
+    #auth.verifier.BasicAuthHeaderAuthVerifier.urls.excludes=/api/liferay/do
 
     #
     # DigestAuthenticationAuthVerifier
@@ -3646,10 +3646,10 @@
     #auth.verifier.DigestAuthenticationAuthVerifier.urls.includes=N/A
 
     #
-    # ParameterAutoLogin
+    # ParameterAuthVerifier
     #
-    #auth.verifier.ParameterAutoLogin.hosts.allowed=
-    #auth.verifier.ParameterAutoLogin.urls.includes=N/A
+    #auth.verifier.ParameterAuthVerifier.hosts.allowed=
+    #auth.verifier.ParameterAuthVerifier.urls.includes=N/A
 
     #
     # PortalSessionAuthVerifier
